Flytter dto-mapping av vedtak til egen fil
